Added slewing to the zoom keys: plus, equals, underscore, and minus.
It was a simple change and is pretty cool.  I had originally intended
to have control+ those keys take larger zoom steps but this was much
easier to implement.  If people don't like this, I'll be glad to
remove it.  I did it to make debugging zoom-dependent label thresholds
easier.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=699649