Don't add labels if the text is empty.  Does not affect the appearance but it can save a little time.
Added code to draw a filled rectangle behind the mouse hover labels.
It's commented out for now because it must wait until KStarsData
becomes a singleton to be implemented so we can easily get the
background color.  This may never be a good idea but it would be fun
to give it a try someday as part of my  no label overlap obsession.


svn path=/trunk/KDE/kdeedu/kstars/; revision=699949