Implemented simple scaling of the star label mag limit by zoom factor. The mag spinner in the setup menu now affects the fully zoomed out label mag limit.  The mag limit is increased as the user zooms in by a factor I chose to keep the label density roughly constant.
On my system when I set the spinner to 3.0 and had no other labels
displaying, the label hit-ratio was roughly 50%.  You can see the
label hit-ratio by pressing the "b" key.  It is the number of labels
drawn divided by the number of draw label requests.

I then scaled the label mag limit as a function of the zoom factor
that I sort of picked out of a hat that seemed to keep the hit ratio
at about 50% as I zoomed in.

I also modified SkyLabeler::stats() so it no longer prints an itemized
list of how many of which type of label got requested which is not very
interesting.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=700161