Experimental: Am now drawing a filled rectangle the color of the sky background behind transient star labels.  I'm not doing it for non-stars ATM so you can compare and contrast the two different ways.  If the object is surrounded by empty space, there is no visual difference.  I noticed the biggest difference when the object was in the Milky Way because then you see the rectangle clearly.
It would be easy and fast to first check to see if the transient label
would overlap an existing label and then only draw the background
rectangle in the case of an overlap.  PLMK if you would like to see
this.  It would be very easy to implement.  IMO it is best to always
draw the rectangle and thus ensure the transient label is always
legible.

Also cleaned up SkyMap::drawTransientLabel() a little bit.

BTW: I don't know why the asteroid transient labels don't exactly
overlap the normal labels.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=700573