Changed the mesh level from 5 to 4.  This seemed to speed things up when the screen was zoomed out.  You are encouraged to try different mesh levels yourself.  You can change it in skymapcomposite.cpp on line 55.  I don't recommend using levels greater than 6.
FYI:

  level    # of trixels
  -----    ------------
    2          128
    3          512
    4        2,048
    5        8,192
    6       32,768

Added 4 new data files that contain pre-computed indices for the
constellations.  This greatly speeds up the constellation boundary
initialization.  If the index file does not exist, we fall back
to computing the indices.  There is a separate index file for
mesh levels 3 through 6.  There is also a mechanism for creating
the data files in ConstellationBoundaryLines::init() by setting
verbose to -1. This will cause the indices to be computed and
printed on stdout.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=700649