Am now generating a KStarsSplash window when re-loading faint stars (which can take many seconds). You can see it be simply increasing the "faint limit zoomed in" in the catalogs setup menu and then hit "okay" or "apply".
I slightly modified KStarsSplash to let the constructor accept an
optional custom message that will replace "Welcome to KStars [...]"
message.

The window is generated from within StarComponent::draw().  The first
time we recognize a need to reload the data, we just create the splash
screen but we don't reload the data.  This allows us to complete the
first draw which will repaint the SkyMap and fill any voids created by
the setup menu window.

I don't trigger a second draw but it always seems to come.  You can
see when because I am printing out to the console "reading data ..."
followed by "Done!" before and after I re-read the data file.

Known bugs:

Sometimes the KStarsSplash window comes up blank.  I think this has to
do with where the mouse is when it opens.  We used to have this some
problem with the original KStarsSplash window.

If we are only loading a few stars, we shouldn't bother displaying the
splash window.

Ideally we would update the splash window with the % of new stars
loaded.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=700965