Further reduced the HTM mesh level from 4 down to 3 which has further improved the drawing speed when zoomed out.  A level-3 mesh has 512 trixels total down from 2048 for level-4 and 8192 for level-5.
Fixed a bug in deciding when to calculate the no-precess aperture
which caused the equator to sometimes not get displayed.

Added timing code in SkyMapComposite::draw() to time the entire draw
cycle.  Uncomment line 293 to get the times to print out:

//kDebug() << QString("draw took %1 ms").arg( t.elapsed() ) << endl;

I discovered that the anti-aliased drawing can take about ten times
longer than the integer drawing.  It is important to turn the
anti-aliasing off in order to test the timing of the rest of the
system.

I can now draw a reasonable skymap, fully zoomed out, with star
labels in about 50 ms.  This increases to about 100 ms when I
include the solar system and asteroid labels.  This then jumps
to about 600 ms if I turn the anti-aliasing on.

CCMAIL:kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=703134