Switching from magnitude spinboxes to slider widgets in the Configure window, for controlling the label density of asteroids and stars.
We used to provide a magnitude limit, and all objects brighter than
the limit would be labeled.  We now have a more advanced labeler that
avoids overlapping text, so a UI control that controls the relative
density of labels is more appropriate.

I updated all code that interacted with the old controls.  The new
parameters are named:
Options::starLabelDensity()
Options::asteroidLabelDensity()
Everything seems to be working as expected.

I wasn't sure what range of values would be useful, so right now the
range is 0 to 20.  This can be changed in the code itself, or in the
*.ui files.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=703448