Reduced the maximum star label density.  If you think it is too small, hit the "b" key and see what the hit ratio is.  If it is 50% or less then the label density is not too small.  If the hit ratio is significantly less than 50% then the label density may even be too high since most labels are not getting drawn because they would overlap.
Prioritized the star labels based on magnitude.  This was not exact.
I bin the labels based on the first two significant digits of the
star magnitude to avoid the expense of a sort.  I could not see any
speed difference using this scheme.  The downside is that the labels
for stars that have magnitudes within 0.1 will be drawn with the same
priority.

This brought the drawing of star labels outside of SkyLabeler::draw()
which then enabled me to bring the drawing of the constellation names
out of that routine as well which cleaned things up a little.  Label
priority is now set by the order things are drawn in
SkyMapComposite::draw() as well as in SkyLabler::draw().  

Cleaned up the starcomponent.h and SkyMapComposite::draw() a little.

CCMAIL:kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=703551