Adding batch modes for the Julian Day and Sidereal Time calculator modules. Before testing, please compile and install the latest ExtDateTime library.
I seriously doubt I am going to have time to implement batch modes for
all of the calc modules in time for 4.0 (feature freeze is in a couple
of days).  Should we leave in the few that have been implemented, or
disable them all for the 4.0 release?

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=704657