Adding KDE4_NO_ENABLE_FINAL, to avoid the compile errors discovered by  Nicolas Lecureuil.
This is just a workaround; I would prefer the htmesh code to be
modified so that it will work under KDE4_ENABLE_FINAL as well.  It may
just be a matter of adding some "if(WIN)" (or similar) directives to
CMakeLists.txt.

It's strange that I can't see how the win-specific code is being
skipped when enable-final is not used...

CCMAIL: kstars-devel@kde.org
CCMAIL: neoclust.kde@gmail.com


svn path=/trunk/KDE/kdeedu/kstars/; revision=706921