Applied the simple fix Jason found for the coordinate grid lock.  He called existing clipping code in the LineListIndex draw loop which keeps us from asking QT to draw the off-screen sections of lines. This is only an issue when anti-aliasing is on but it is always used (for line segments not already clipped by the celestial horizon).  The addition of the clipping code had no measurable impact on the speed. With anti-aliasing turned off the coordinate grid never takes more than a few milliseconds to draw but with anti-aliasing turned on it can take hundreds of milliseconds.
Fixed a minor bug in onscreenLine() that caused lines near the bottom
edge to disappear.

Am now always using the index for components that don't precess.
Previously, I had disabled the no-precess index for very low zooms in
order to save the time of computing a 2nd aperture but the aperture
only takes a millisecond or two (with the level-3 mesh) so we always
use it now.

Removed most skycomponents code that was switching to integer drawing
when anti-aliasing was off.  The anti-aliasing is totally controlled
by QPainter::setRenderHint() so that code added complication but didn't
save us any time.  The number of draw routines in LineListIndex was
thus cut in half from six down to three.  And since we are now always
using the no-precess index, the drawAllLines() is not being used.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=708070