Made a bool SkyMap:onscreenLine2() routine that uses the return value to flag if a line is entirely offscreen instead of changing the values of the points.
Fixed a minor bug in onscreenLine2() where a line would disappear if
one endpoint of the line was barely offscreen.  This bug probably
still exists in onscreenLine().

Varied the point density of the coordinate grid lines of constant
declination so the lines near the poles have fewer points.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/KDE/kdeedu/kstars/; revision=708436