Printing the sky map now works.
This is a pretty extensive change, but almost all of it was a simple
refactoring.  Namely, instead of passing a "scale" argument to dozens
of draw-related functions, the scale is now a member variable of SkyMap.

As a reminder, the scale factor is always 1.0, unless we are printing.
Then the scale factor is determined by the properties of the printer.
Essentially, it's the ratio of the width (or height) of the printer's
QPaintDevice to that of the SkyMap widget.

This is a nice change, because few people knew what scale was for, so
now we don't have to pass a mysterious argument all over the codebase.
Most of the changes in this commit involve the removal of the scale
argument from various functions.

Also, I am using QPrinter and QPrintDialog instead of KPrinter,
because from what I've read, KPrinter will probably not be ready for
KDE-4.0.

CCMAIL: kstars-devel@kde.org



svn path=/trunk/KDE/kdeedu/kstars/; revision=722230