Fixing bug #151206: updating ephemerides for comets and asteroids.
The new ephemerides are also available via the "Get New Stuff" action.
I am not porting to the 3.5 Branch, because it would break the strings
freeze.  3.5 users can simply use the GNS action.

TODO: make the file-parsing code more robust so that the comets dat
file can be used as published by JPL, without formatting changes.
This used to be the case, but JPL changed the widths of their columns,
which breaks ouur fragile column-based parser.

BUG:151206
CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=730537