Usability fix for the "Start/Stop clock" action.
In 3.x, we used a custom "ToggleAction" class that allowed us to
change the icon of the action from Play to Pause when it was toggled.
Some people found this confusing (does the Play icon mean it's
currently playing, or that it will play if you click it?).

In trunk, the icon toggling hasn't been working, so it's been even
worse: the icon is always "Pause", with no indication of what state
it's in.

Now, the icon is still always Pause, but we are using KToggleAction,
so that when the clock is paused, the button has a "pressed" look.
So, you press the button to pause the clock, and un-press it to start
it again.

CCMAIL: kstars-devel@kde.org



svn path=/trunk/KDE/kdeedu/kstars/; revision=730760