Fix some issues with the Links tab of the Details dialog.
Using any of the action buttons (View, Edit, Remove) would always act
on the item in the top list (the list of webpages).  The desired
behavior is that it acts on the last-clicked item.

Now storing the last-clicked item as a member variable, rather than
trying to keeep track of which listwidget was last clicked.

Also, the edit link dialog now allows you to change both the link URL
and the text describing the link (before, only the URL could be
changed).

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=738311