Fixing bug #152869
Hidden Toolbars not persistent on restart.  Also fixed the same issue
for the statusbar.  I had thought that calling setChecked() on the
KToggleActions would trigger the connected slotShowGUIItem(), but
apparently not.

Interestingly, the call to slotGUIItem() *is* successfully made in the
trunk codebase...so no fix needed there.

Thanks for the report!

BUG: 152869
CCMAIL: kstars-devel@kde.org


svn path=/branches/KDE/3.5/kdeedu/kstars/; revision=741363