Initialise province variable with empty string
When I was searching for "ljubljana" in KStars wizard the city couldn't
be found. Before I started entering the name into city name field the
city was on the list of all cities. After I started entering letters of
the city name it suddenly just dissapeared from the list. I checked it
closely and it appeared that all locations which didn't have province
specified disappeared from the list after the filter kicked in. This
fixes the problem for me.

svn path=/trunk/KDE/kdeedu/kstars/; revision=744979