Fix drawing of the opaque ground for the Equirectangular projection. When in Alt/Az mode, Equirectanguar is a simple case, because the ground polygon is always just a rectangle.  To accomodate this special case, I needed to break the label-drawing code out into its own function: HorizonComponent::drawLabels().
Also removed integer code from HorizonComponent::draw() for
non-antialiased drawing.  We don't need separate code for this, the
painter can handle float values in integer mode.

Also made the Horizon the last thing to be drawn in
SkyMapComposite::draw(), so that no labels are visible "through" the
opaque ground.

TODO: Because the horizon is now being drawn last, its label and the
compass labels have lower priority than other labels.  This is fine,
except it looks strange when the opaque ground is drawn, because the
labels sometimes disappear for no apparent reason.  When the opaque
ground is drawn, the horizon labels should not be considered to
overlap with sky labels.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=749120