Remove the KStars pointer argument from all draw functions in the skycomponents classes.  This pointer was *only* used to reference the SkyMap and KStarsData pointers, but since these classes are singletons, we can reference them through the static Instance() function.
A similar cleanup needs to be done for the init() and update()
functions (which all take an unnecessary KStarsData* argument), but I
think I'll wait on that until after the freeze.

I'm doing this fix now because it resolves a crash condition when
using image-dump mode.  Despite the large number of files touched,
this is really a trivial change.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=750936