ScriptBuilder fixes:
Show time-reated functions under a separate heading (SimClock) in
the Function Browser.

Show functions in the Function Browser in alphabetical order

The OptionsTreeView dialog lists all of the view options that can
be modified by changeViewOption(), along with a text description
of each and the data type of its argument.  Fixed this dialog to
show the option names, not just their descriptions.

The previous fix also fixes the problem that the selected view
option did not get propagated to the ScriptBuilder tool when the
OptionsTreeView was closed.

Resize the OptionsTreeView window, and the columns in the
treeview, to show most of the text in the name and description
columns.

Do not allow the user to edit thetext in the FuncDocs pane (where
API information about each function is displayed).

TODO: Fix the logic used to determine under what conditions the
Save/Overwrite warning messageboxes should be shown.  Quite a few
problems hee...

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=751089