Applying Akarsh's patch to fix calculation of asteroid magnitudes. I only changed "10" to "SkyObject::ASTEROID" where appropriate.
In testing the magnitudes, I found that the KStars values are always 
fainter than the published NASA values (at ssd.jpl.nasa.gov) by a few 
tenths of a magnitude.  I believe this is because the geocentric 
distance computed by kstars is larger than the NASA values.  I'll have 
to look into why that is happening.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=751453