More ScriptBuilder fixes:
Don't ask for confirmation when saving a script that had been
opened or previously saved.

Parse qouted strings in arguments of script functions properly
(someone had wrapped this code in "#if 0", so please verify)

Don't disable the "Save As" button after a save.  "Save" gets
disabled to signal that no unsaved changes have been made.
The "Save As" button should be enabled as long as the current
script is not empty.

Don't signal unsaved changes if a function is simply selected in
the function list

When the last function in the script is removed, make the
previous function the selected one.

Update the argument widget whenever the selected function changes.

Discard the script when the scriptbuilder window is closed.

make sure string arguments are quoted when writing the script

Don't show "I18n_EMPTY_MESSAGE" in province field of the
setGeoLocation argument widget, if the province of the selected
city is undefined (just leave it blank)

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=751792