Applying patch to improve rendering time for stars by a factor of twenty.  Stars are now drawn as pixmaps, rather than calling drawEllipse() for each star.  The star pixmaps are stored in a static QHash in StarObject.  The hash contains images for all spectral types, and all sizes.  This is how stars were rendered in 3,5, so this is a regression fix.
The star pixmaps are regenerated when the user selects a new color
saturation level for stars (this is effected by modifying the width of
the QPen used to draw the colored rim of the circle), and when the
user chooses a color scheme that uses solid red, black or white star
images (without a colored rim).

Also, stars are now rendered slightly larger on-screen, closer to what
was done in 3.5.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=752585