The INDI network backend was migrated to fully use Qt, this along with changes in the INDI parser led to orders of magnitudes improvement in processing and constructing on the fly devices. Binary BLOBs gained the most advantage, loading an average FITs file from the CCD now takes 20% of what is used to take.
It is now possible to start multiple devices managed by one INDI server vs the one-device-one-server behavior before. This model has also resulted in unifying server and local devices with remote devices and that in turn greatly simplified the code. Now that this is the 4.1 branch, there is quite a bit of refactoring to improve consistency. 

CCMAIL:kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=757384