Second attempt to fix bug #154195.  Thanks for continued testing, please try this revision as well.  I think/hope I have it this time.
I am using QString::toDouble() instead of QTextStream::operator>>
The API docs for toDouble specifically say it uses the local 
decimal setting, and if that doesn't work, it falls back to use 
".", no matter what the locale uses for decimalSymbol().  Because
of this fallback, I am not converting "." to decimalSymbol() anymore.
In theory, it should still work now that I am using QString::toDouble().

CCBUG: 154195


svn path=/trunk/KDE/kdeedu/kstars/; revision=764110