Cosmetic fixes for different projection schemes. The opaque ground polygon is now well-behaved for all projections.  Fixed a crash at low zoom when using Gnomonic projection, as reported by Mederic Boquien.
Gnomonic is a special case, because unlike the other 
projections, it has an infinite "sky horizon" 
boundary.  i.e., the locus of points that are 90 
degrees from the focus is projeced as a circle in 
other systems, but these points are at infinity in 
the Gnominic projection.

Also forward-porting the fix for Bug #133505 (crash 
condition in SkyMap::refract()).  Will backport these 
changes to the 4.0 branch as well.

CCMAIL: boquien@astro.umass.edu
CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=766933