Reorganizing initFocus():
+ Fix bug #162059: when leaving the program in Horizontal coordinates 
and not tracking on the sky, the initial position on next startup should 
be the same Alt/Az coordinates.  Fixed.

+ When leaving the program tracking on a star named "star", the initial 
position on the next startup should be that same star,  Fixed.

They seem like simple fixes, but it actually took quite a bit to fix 
these.  Another symptom of the bubble-gum-and-duct-tape nature of the 
startup code...

Is it worth trying to backport this fix to the 3.x codebase?  Are we 
going to have 3.5.10?

CCMAIL: kstars-devel@kde.org
BUG: 162059


svn path=/trunk/KDE/kdeedu/kstars/; revision=809477