We have 6 "object action" keystrokes that operate on the clickedObject: C = center the object D = Show the object's Details window P = Show the object's popup menu O = Add the object to the observing list L = Add/remove the object's name label T = Add/remove a trail from the object (if isSolarSystem()==true)
If the Shift key is held, then these actions operate on the *centered 
object (i.e., focusObject) instead.  This is accomplished by setting the 
clicked object to be the focus object (setClickedObject( focusObject() 
)).  With this commit, the re-assignment of clickedObject when Shift is 
used with an action key is now temporary; it now reverts back to 
whatever it was before the action was invoked.

CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=815443