Implementing dynamic loading of star data from harddisk, depending on current trixel.
At (zoom out mag limit, zoom in mag limit) = (6.3, 10.0) with
sufficient zoom (I couldn't measure it, but I could expect not more
than 6 trixels to be present), we get a saving of about 30kB in
memory.

This is still bad because we are allocation 179 StarBlocks, so it
looks like we still have some bug.

CCMAIL: kstars-devel@kde.org


svn path=/branches/kstars/summer/kdeedu/kstars/; revision=819780