Added large zoom steps to unfrozen branch.
The change in zoom due to the mouse wheel or "-" and "=" keys is now 2.0
instead of the previous 1.1.  This reduces the number of key presses or
wheel clicks to change to to a given zoom by roughly a factor of eight.

The Shift and Control keys reduce the zoom step by factors of 2 and 4
respectively.  Combined, they reduce it by a factor of 8 which is close
to (but not equal to) the original factor of 1.1.

Also increased MINZOOM from 200 to 250, which I think roughly compensates
for the change caused when we converted SkyMap::fov() to reporting the
diagonal of the screen instead of the minimum of the height and width.
The change in fov() was made to provide a radius for the SkyMesh::aperture()
call.


CCMAIL: kstars-devel@kde.org

svn path=/branches/kstars/unfrozen/kstars/; revision=823176