Presto, chango! The statusbar now has a (non-functional) KLed on its right side.  If you want it on the left, move the new lines to the top of initStatusBar().
The public accessor is KLed* KStars::diskLED()
The relevant functions to call are diskLED()->on() and 
diskLed()->off() (or diskLED()->toggle()).

(BTW, FYI: this is going to be used to indicate when the 
disk is being accessed to dynamically load more stars.  
It will probably be configured to only show up for debug 
builds, or at least will be hidden by default)

CCMAIL: kstars-devel@kde.org


svn path=/branches/kstars/unfrozen/kstars/; revision=824175