kludgy-but-should-work fix for Bug #165170 (Cunjunctions not calculated for Japanese locale.
Thank you for reporting this localization bug...please report any other problems 
related to localization that you may run into!  I did see your second report as 
well, and I will get a fix ASAP.

The problem is that ComboBox->currentText() is translated, but the KSPlanet ctor 
expects the untranslated name string as a parameter.

This solution uses currentIndex() instead, and a switch to determine which planet 
should be used.  The solution is a bit fragile, because the planets are now 
assumed to appear in a particular order in the combobox. 

I am planning to implement a better solution in the unfrozen branch, so please 
don't port these changes to unfrozen or summer; wait for the unfrozen fix.

CCMAIL: kstars-devel@kde.org
BUG: 165170


svn path=/trunk/KDE/kdeedu/kstars/; revision=825981