Welcome the Tycho 2 catalog with stars down to mag 12.5 consistently, and some stars fainter than that.
I hardly notice any memory overhead with sensible magLimitZoomOut
(5.5) and memUsage (9) settings. This is probably because the maximum
memory requirement is not at the maximum zoom, but at intermediate
zoom levels when maglim = 9.5 or so.

Thanks a lot for the catalogs Jason!

CCMAIL: kstars-devel@kde.org


svn path=/branches/kstars/summer/kdeedu/kstars/; revision=826096

The following changes were in SVN, but were removed from git:

M	kstars/data/deepstars.dat