porting fix from trunk, revision 822804.  Quoting the original commit message:
Fixing a usability issue in the options window.
Toggling the view toolbar buttons now toggles the
corresponding control in the options dialog.  This is
useful for those who keep the options window open while
using the program.

TODO: there's no meta-option for deep-sky objects in the
options window.  I'll ask translators if it's ok to add
one for 4.2.

Thanks for pointing this out, James.

CCMAIL: kstars-devel@kde.org


svn path=/branches/kstars/unfrozen/kstars/; revision=826610