One more port from trunk, revision 826203.  Quoting original commit message:
Forcing the algorithm to find conjunctions take finer steps when
dealing with the Moon, Venus and Mars, than before.

Also removing the condition on prevSign to check for a minimum. This
condition was probably incorrect, and removing it seems to produce
more (correct) results.

These fixes certainly make the conjunction tool slower, but more
exhaustive in listing out conjunctions.

CCMAIL: kstars-devel@kde.org


svn path=/branches/kstars/unfrozen/kstars/; revision=826613