Applying variable timestep patch from trunk (r827206).
Akarsh, your commit message mentioned fixing a few bugs that my original patch 
revealed...was there anything besides reducing the tolerance of findPrecise() to 1 
minute?

Also adding a nice progress bar that appears in place of the Compute button while 
the computation takes place, and then disappears when it's done.

TODO: 

+ it would be nice to see conjunctions added to the results box as they are found, 
rather than waiting until the computation is finished.

+ Add a "Show conjunction" button that will set the date/time to the time found by 
the tool for the highlighted event, and set the focus to Object1.

CCMAIL: kstars-devel@kde.org


svn path=/branches/kstars/unfrozen/kstars/; revision=827475