Remove KStarsData pointers from function argument lists (use KStarsData::Instance() instead).
Stage 1: SkyComponents.

I have added a protected KStarsData* member to SkyComponent (so all 
Components/Composites inherit this member).  The pointer is assigned to 
KStarsData::Instance() in the component's init() function.

Question: does it make sense to store this pointer as a member, or should we 
simply call KStarsData::Instance() locally in the functions that need it?

Note: I made init() pure virtual in abstract classes (i.e., "virtual void init() = 
0;" instead of "virtual void init() {}") to make sure they can't be instantiated.

As a result of this change, I had to add a CustomCatalogComposite class, because I 
had been using a SkyComposite directly to store our custom catalog components, and 
I decided that SkyComposite should be an abstract class.

CCMAIL: kstars-devel@kde.org


svn path=/branches/kstars/unfrozen/kstars/; revision=829109