Fixing bug #166002 (Option to save DSS image).
The ImageViewer dialog has Save and Close buttons, but due to a layout problem, 
they were not visible in the window.  I added a QFrame container for the image 
canvas and the text caption, and made this container the dialog's mainWidget.  
This makes the dialog buttons visible again.  

Unfortunately, it also causes a pretty thick margin to be added around the 
image/caption.  Maybe a designer file for the dialog could solve this issue, but 
at least the Save button is accessible again.  Too bad it's too late for 4.1...

Thank you very much for the report Prakash!

BUG: 166002
CCMAIL: kstars-devel@kde.org


svn path=/branches/kstars/unfrozen/kstars/; revision=829315