Backporting fix for FOV symbols (trunk r848577).  Original commit message fllows:
Applying patch from Mederic Boquien.  Quoting his message:

"I have made a patch to correct a few problems on the FOV tool:
* It was not possible to use the semitransparent FOV because in kstars.kcfg,
the max index was set to 3 whereas the semitransparent FOV had index #4. When
selecting a semitransparent FOV, it switched back to #3 (bull's eye)
* Now the FOV preview is anti-aliased and the text indicating the FOV size is
correctly centred
* Now the semitransparent FOV is really semitransparent using the alpha
channel and not filled with a not-so-pretty pattern
* After having edited a FOV, the preview in the dialog containing the list of
FOV is updated"

Thanks, Mederic!



svn path=/branches/KDE/4.1/kdeedu/kstars/; revision=850201