Fixing bug #169744 (Catalog fields can't be modified because drag events not working).
Well, the drag events work now, but it's difficult to predict what the order of 
items will be when you drop an item.  It's supposed to insert the item in the 
correct place based on the cursor's position, but it looks like there are some 
problems.  I'll leave the bug open until this is working as well as it did in 
3.5.x...

CCBUG: 169744
CCMAIL: kstars-devel@kde.org


svn path=/trunk/KDE/kdeedu/kstars/; revision=851910